change accountid from int64 to int32

+++ server/api/transactions.pb.gw.go
@@ -84,7 +84,7 @@ func request_TransactionsService_List_0(ctx context.Context, marshaler runtime.M
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "missing parameter %s", "account_id")
 	}
 
-	protoReq.AccountId, err = runtime.Int64(val)
+	protoReq.AccountId, err = runtime.Int32(val)
 
 	if err != nil {
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "type mismatch, parameter: %s, error: %v", "account_id", err)
@@ -118,7 +118,7 @@ func local_request_TransactionsService_List_0(ctx context.Context, marshaler run
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "missing parameter %s", "account_id")
 	}
 
-	protoReq.AccountId, err = runtime.Int64(val)
+	protoReq.AccountId, err = runtime.Int32(val)
 
 	if err != nil {
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "type mismatch, parameter: %s, error: %v", "account_id", err)
